tests/wpt/tests/css/css-shapes/shape-outside-invalid-ellipse-001.html

<!doctype html>
<html>
<head>
    <meta charset="UTF-8">
    <meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if ellipse shape function has two invalid arguments">
    <title>CSS Shapes Test: Ellipse - invalid 2 arg</title>
    <link rel="author" title="Sherlock Zhong" href="mailto:sherlock.z0917@gmail.com">
    <link rel="reviewer" title="Alan Stearns" href="mailto:stearns@adobe.com"> <!-- 11-09-2013 TestTWF Shenzhen -->
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-ellipse">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
#shape{
    width: 300px;
    height: 300px;
    shape-outside: ellipse(50, 50);
}
    </style>
</head>
<body>

    <div id="log"></div>

    <div id="shape"></div>

    <script>
    var obj = document.getElementById('shape');
    var val = getComputedStyle(obj).shapeOutside;
    test(function() {assert_equals(val, "none", "declaration should not be valid")});
    </script>

</body>
</html>
